Actualizado y probado con time
No usen mas de 8000 o se les pasmara la compu

+++ 5_static_STL_ver2/PoligoIrregular.cpp
@@ -21,8 +21,8 @@ void PoligonoIrreg::printVertice() {
 
 }
 
-int PoligonoIrreg::vecesMuestF = 0;
+double PoligonoIrreg::vecesMuestF = 0;
 
-int PoligonoIrreg::getVecesMuestF() {
+double PoligonoIrreg::getVecesMuestF() {
     return vecesMuestF;
 }
